requirements.txt
numpy==1.24.3
pandas==2.2.3
geopandas==1.0.1
shapely==2.0.7
scikit-learn==1.2.0
h3pandas==0.2.6
h3==3.7.7
python-geohash==0.8.5
tqdm==4.65.0
torch==2.0.1
transformers==4.29.2

scipy
polygeohasher